Create p2.py
Create a line plot to visualize a simple linear relationship between two sets of numeric data. 
Specifically, plot the sequence of numbers `[1, 2, 3, 4, 5, 6]` on the x-axis and their corresponding 
multiples of 5 `[5, 10, 15, 20, 25, 30]` on the y-axis. The x-axis should be labelled "Sequence Number" 
and the y-axis should be labelled "Multiples of Five". Use the `matplotlib` library to generate and 
display this plot, clearly illustrating the trend of the multiples of five as the sequence number 
increases.