sample 5
 Create a line plot to visualize the relationship between two sets of non-uniform values. Plot the 
values [6, 2, 4, 8, 1, 5] on the x-axis and the values [2, 4, 9, 7, 3, 9] on the y-axis. Label the x-axis as 
"Random Number set 1" and the y-axis as "Random Number set 2". Use the matplotlib library to generate and display the plot, illustrating how the values in the second set change with respect to the 
values in the first set.